alterações gerais

+++ Augustus Fashion/Model/Testes.cs
@@ -32,7 +32,7 @@ namespace Augustus_Fashion.Model
         public static bool ValidarCelular(this string valor) =>
             new Regex(@"^[0-9]{11}$").Match(valor).Success;
 
-        public static bool ValidarDesconto(this float valor) =>
+        public static bool ValidarDesconto(this decimal valor) =>
             new Regex(@"^[0-9]{1,2}$").Match(Convert.ToString(valor)).Success;
     }
 }
